pokazuje markery

+++ FriendsZone/FriendsZone/Models/Spot.cs
@@ -11,8 +11,8 @@ namespace FriendsZone.Models
         public int id { get; set; }
         public int gid { get; set; }
         public string name { get; set; }
-        public decimal latitutde { get; set; } 
-        public decimal longitude { get; set; }
+        public double latitutde { get; set; } 
+        public double longitude { get; set; }
         public string description { get; set; }
 
         public override string ToString()
